Add initial module version
Add an initial version of the module which overrides the
AzureBlobStorageProvider and changes the provided GetPublicUrl method used
by media to direct to a new route. This new route is handled by
MediaController which checks for a newly provided permission and serves up
the file if it is available. This means the media can be served from a
private azure blob storage container, allowing secure access to assets.