Add SecureMediaPart
Add the SecureMediaPart which attaches an IsSecure boolean to Media
items.

Add a migration which will attach the SecureMediaPart to all items with
the 'Media' stereotype except for ones which contain 'external' in the
name.

Add Driver and edit view for the SecureMediaPart.

Add placement info allowing placement of the SecureMediaPart.

Update the name of the feature again and make the whole module a single
feature instead of having a sub feature for the media filtering.